--- reviews/2022/a.md
@@ -1,16 +1,0 @@
----
-book:
-  author: A
-  narrator: A
-  cover:
-    name: a.jpg
-    size: 151115
-    tint_color: "#ce4339"
-  publication_year: "2021"
-  title: A
-review:
-  date_read: 2022-03-13
-  format: audiobook
-  rating: 4
----
-
